Rearange structure of this projec to not expose a public API
Some parts of this project may be useful to other cargo projects, but
if we expose an API we should document it properly and also make some
concious commitments to its stability. Thus, for now I removed the
`lib` part completly, leaving this repository as `bin`-only.

Furthermore, cleaned up (simplified) some thing in the assambly of the
inferior command.