Fix parsing bug when argument has quotes
Because formerly we split on whitespace and tested for equality with our
"special" argument strings, an argument with quotes would not be
properly parsed and replaced.

Rather than checking for equality with a single string, we use
`contains` and `replace` while treating the arguments as a plain string.
This means that we will always replace the special argument sequences
even though they are surrounded in quotes.

The solution is rather verbose and does a lot of allocations, however as
a commandline tool this will be negligible.

Fixes: #20